+++ .eslintignore
@@ -1,5 +1,6 @@
 locales/locales.ts
-ts/utils/__tests__/xss.test.ts
+**/__tests__/**
+.eslintrc.js
 definitions/*
 src/react-app-env.d.ts
 Dangerfile.ts
