+++ package.json
@@ -68,5 +68,11 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-sonarjs": "^0.10.0",
     "prettier": "^2.4.1"
-  }
+  } ,
+  "jest": {
+    "coveragePathIgnorePatterns" : [
+      "src/index.tsx",
+      "src/reportWebVitals.ts"
+      ]
+    }
 }
